+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixes
 * UiTextbox and UiAutocomplete: Fix `debounce` prop handling
 * UiFab and UiIconButton: Set icon width to `100%` to fix horizontal centering in Safari on macOS
+* UiButton and UiIconButton: Fix `dropdownPosition` prop not being passed to UiPopover
 
 ### Breaking
 * UiTextbox: `autocomplete` prop type changed to String, to allow for configuring browser autocomplete